Refactor overlord state in its own package. Move state interface where it's used. [Note this does not compile yet, followup CL on interests comng up]
