+++ src/go/overlord/BUILD.bazel
@@ -2,18 +2,15 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "overlord.go",
-        "statemanager.go",
-    ],
+    srcs = ["overlord.go"],
     importpath = "github.com/google/minions/go/overlord",
     visibility = ["//:__subpackages__"],
     deps = [
         "//go/overlord/interests:go_default_library",
+        "//go/overlord/state:go_default_library",
         "//proto/minions:go_default_library",
         "//proto/overlord:go_default_library",
         "@com_github_google_uuid//:go_default_library",
-        "@com_github_patrickmn_go-cache//:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_x_net//context:go_default_library",
     ],
