+++ src/go/overlord/state/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["statemanager.go"],
+    importpath = "github.com/google/minions/go/overlord/state",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//proto/minions:go_default_library",
+        "//proto/overlord:go_default_library",
+        "@com_github_patrickmn_go-cache//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["statemanager_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//proto/minions:go_default_library",
+        "//proto/overlord:go_default_library",
+        "@com_github_stretchr_testify//require:go_default_library",
+    ],
+)
